Refacto timelines
