+++ package.json
@@ -256,7 +256,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-push": "yarn lint && yarn test:e2e && yarn build"
+      "pre-push": "yarn lint && yarn build"
     }
   }
 }
