--- src/components/icons/ghost-timeline-bullet-point/ghost-timeline-bullet-point.stories.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-import { GhostTimelineBulletPoint } from "./ghost-timeline-bullet-point";
-
-export default {
-  title: "icons/Timeline Bullet Point",
-  component: GhostTimelineBulletPoint,
-};
-
-export const StandardGhostTimelineBulletPoint = (): JSX.Element => {
-  return <GhostTimelineBulletPoint />;
-};
