--- src/components/icons/ghost-timeline-bullet-point/ghost-timeline-bullet-point.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const GhostTimelineBulletPoint: React.FC = () => {
-  return (
-    <svg width="8" height="8" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <circle cx="4" cy="4" r="3.5" fill="#fff" stroke="#F0F1F3" />
-    </svg>
-  );
-};
