+++ src/components/icons/github-icon/github-icon.stories.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { GithubIcon } from "./github-icon";
 
 export default {
-  title: "icons/social/Github Icon",
+  title: "icons/Social/Github Icon",
   component: GithubIcon,
 };
