--- src/components/icons/timeline-bullet-point/timeline-bullet-point.stories.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-import { TimelineBulletPoint } from "./timeline-bullet-point";
-
-export default {
-  title: "icons/Timeline Bullet Point",
-  component: TimelineBulletPoint,
-};
-
-export const StandardTimelineBulletPoint = (): JSX.Element => {
-  return <TimelineBulletPoint />;
-};
