--- src/components/icons/timeline-bullet-point/timeline-bullet-point.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const TimelineBulletPoint: React.FC = () => {
-  return (
-    <svg width="8" height="8" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <circle cx="4" cy="4" r="4" fill="#F0F1F3" />
-    </svg>
-  );
-};
