+++ src/components/icons/vkontakte-icon/vkontakte-icon.stories.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { VKontakteIcon } from "./vkontakte-icon";
 
 export default {
-  title: "icons/social/VKontakte Icon",
+  title: "icons/Social/VKontakte Icon",
   component: VKontakteIcon,
 };
