--- src/components/timeline-end/timeline-end.stories.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-
-import { StoriesContainer } from "../containers/stories-container";
-import { TimelineEnd } from "./timeline-end";
-
-export default {
-  title: "components/TimelineEnd",
-  component: TimelineEnd,
-};
-
-export const StandardTimeLineEnd = (): JSX.Element => {
-  return (
-    <StoriesContainer>
-      <TimelineEnd />
-    </StoriesContainer>
-  );
-};
