+++ src/components/timeline/timeline.stories.tsx
@@ -1,17 +1,25 @@
 import React from "react";
 
 import { StoriesContainer } from "../containers/stories-container";
-import { Timeline } from "./timeline";
+import { Timeline, TimelineEnd } from "./timeline";
 
 export default {
-  title: "components/Timeline",
+  title: "components/Timelines",
   component: Timeline,
 };
 
-export const StandardTimeLine = (): JSX.Element => {
+export const TimeLine = (): JSX.Element => {
   return (
     <StoriesContainer>
       <Timeline />
+    </StoriesContainer>
+  );
+};
+
+export const TimeLineEnd = (): JSX.Element => {
+  return (
+    <StoriesContainer>
+      <TimelineEnd />
     </StoriesContainer>
   );
 };
